+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 Table of Contents
 
+- [Unreleased](#unreleased)
 - [v0.2.3](#v023)
     - [Added](#added)
 - [v0.2.2](#v022)
@@ -17,6 +18,10 @@ Table of Contents
     - [Added](#added-1)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
+
+## Unreleased
+
++ content.scale API documented
 
 ## v0.2.3